requirements.txt
pinecone-client
streamlit
tiktoken
requests
